requirements.txt
add stuff here
openai
python-dotenv